LayoutTests/imported/w3c/web-platform-tests/scroll-animations/animation-trigger/timeline-trigger-toggle.tentative-expected.txt

Source
Target
Toggle timeline-trigger

FAIL Toggling trigger on starts controlling attached animation; toggling off stops assert_equals: expected (string) "none" but got (undefined) undefined
